+++ docs/CHANGELOG.md
@@ -48,7 +48,9 @@ Version X
 - Lower minimum character count for console tab complete to 1
 - Add `$DF GunGame` option in dedicated server config
 - Fixed jeep cockpit mesh not being rendered for any jeeps entered after the first in a level
-- Add framework for custom events
+- Add framework in game and editor for custom events
+- Add custom events `Clone_Entity`, `Set_Player_World_Collide`, and `Switch_Random`
+- Remove `You must rebuild geometry before X` warnings in editor